feat(app router confirmation page): handle story and cart data
