+++ apps/store/src/graphql/ShopSessionOutcomeId.graphql
@@ -1,5 +1,6 @@
 query ShopSessionOutcomeId($shopSessionId: UUID!) {
   shopSession(id: $shopSessionId) {
+    id
     outcome {
       id
     }
